librem_14/board_battery: Fix CELL pin setting for charge controller
The BAT_CELL_SEL line (GPIO_J3) needs to be set high and stay high
in order for the charge controller to be set properly depending on
if a 3-cell or 4-cell battery is installed.

Previously, we tried to set the GPIO properly based on the detected
battery type, but since the factory chose the wrong MOSFETs, this
approach did not work; to the contrary, it interferes with the resistor
modification. By rolling this back to pre-1.6 code, everything works
again.

We do have a very slight corner case, where directly after power on (and
before the EC firmware sets new value)s the charge controller assumes
the wrong values for the 4-cell batteries (IIRC). This should not be a
huge problem since we very soon during EC init reset these values
according to the battery probe.

Signed-off-by: Nicole Faerber <nicole.faerber@puri.sm>